% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd.R
\name{alpha_hat}
\alias{alpha_hat}
\title{Estimates the joint alpha for an entire data matrix}
\usage{
alpha_hat(data, k)
}
\arguments{
\item{data}{the \code{matrix} of observations to be used to estimate alpha}

\item{k}{the number of large order statistics to be used for each variable in
\code{data} (this value must be a scalar).}
}
\value{
a scalar estimate of the joint alpha rounded to four decimal places.
}
\description{
This function computes the hill estimate for each column in the \code{data} and
then takes the \code{mean()} because \code{k} is shared. This is the estimate for
alpha under the assumption that all of the variable have the same alpha.
}
\examples{
alpha_hat(financial_data[,-1], k = 300)
}
