% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd.R
\name{alpha_plot}
\alias{alpha_plot}
\title{Create a plot of confidence intervals for alpha from each variable (station)}
\usage{
alpha_plot(data, k, ci = 0.95)
}
\arguments{
\item{data}{the \code{matrix} of observations with variables (e.g., stations) in
the columns.}

\item{k}{the number of largest order statistics to be used with the Hill
estimator}

\item{ci}{the confidence level (as a proportion) for the confidence intervals
that will be plotted (the default is 0.95 which makes 95\% CIs).}
}
\value{
a ggplot2 \code{plot} object which shows 100*\code{ci}\% confidence intervals
for the tail index (alpha) of each variable computed using the Hill
estimator as well as an estimate of the shared (joint) alpha under the
assumption that the alpha is the same for each variable.
}
\description{
The alpha (index of regular variation) for each variable is estimated with
the Hill estimator with a user specified \code{k}. In addition, the \code{ci}\%
confidence intervals are computed. These are plotted with the a horizontal
line which is the estimate for a single joint alpha. The joint alpha is
computed as the average of the individual estimates because \code{k} is the
same for each variable. If a user wants \code{k} to differ then this tool is
inappropriate.
}
\examples{
alpha_plot(financial_data[, -1], k = 300)
}
\references{
\insertRef{hill1975}{tpdmethods}
}
