% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{ar1_tpdf}
\alias{ar1_tpdf}
\title{Compute the TPD function from a TL-AR(1) parameter}
\usage{
ar1_tpdf(phi, max_lag = 20)
}
\arguments{
\item{phi}{the TL-AR(1) parameter}

\item{max_lag}{the TPDF is computed for lags 0 through\code{max_lag}}
}
\value{
a \code{vector} of model TPDF values
}
\description{
This function takes a parameter scalar \code{phi} as input and returns the
model TPDF for lags 1 through \code{max_lag}
}
\examples{
out <- ar1_tpdf(0.3)
}
\references{
\insertRef{mhatre2024arma}{tpdmethods}
}
