% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{arma11_tpdf}
\alias{arma11_tpdf}
\title{Compute the TPD function from a TL-ARMA(1,1) parameter \code{vector}}
\usage{
arma11_tpdf(params, max_lag = 20)
}
\arguments{
\item{params}{a vector of the form (\code{theta}, \code{phi}) where \code{theta} is the
TL-MA coefficient and \code{phi} is the TL-AR coefficient}

\item{max_lag}{the TPDF is computed for lags 0 through\code{max_lag}}
}
\value{
a \code{vector} of model TPDF values
}
\description{
This function takes a parameter \code{vector} as input and returns the
model TPDF for lags 1 through \code{max_lag}
}
\examples{
out <- arma11_tpdf(c(0.9, 0.1), max_lag = 15)
}
\references{
\insertRef{mhatre2024arma}{tpdmethods}
}
