% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_lin_operations.R
\name{f}
\alias{f}
\title{Transformed-linear transformation function}
\usage{
f(y)
}
\arguments{
\item{y}{real number to be transformed}
}
\value{
a scalar on the positive half-line
}
\description{
This function is the softplus function used to map the reals to the positive
half-line.
}
\examples{
f(2.5)
f(-1.2)
}
