% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_lin_operations.R
\name{finv}
\alias{finv}
\title{Transformed-linear inverse transformation}
\usage{
finv(x)
}
\arguments{
\item{x}{a non-negative number}
}
\value{
a real number
}
\description{
This function is the inverse of the softplus (\code{f}). It maps the positive
half-line to the reals.
}
\examples{
finv(2.5)
# finv(-0.1) # this should return an error.
}
