% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire_weather_present}
\alias{fire_weather_present}
\title{Fire Weather Index Data - Present period}
\source{
\url{https://confluence.ecmwf.int/display/CKB/ERA5\%3A+data+documentation}
}
\description{
Fire Weather Index (FWI) data after marginal transformation to Fr\'echet(2)
margins. These data are from the gridbox that contains the Harbison
Meadow RAWS weather station near Grand Lake, Colorado. They are data from
June through October of 2002-2021. Weather data were downloaded from the
ERA5 data on the Copernicus Climate Change Service. A \code{matrix} with 153
rows (days of the fire season) and 20 columns (years).
}
\references{
\insertRef{wixson2023}{tpdmethods}
}
