% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{gen_ar1}
\alias{gen_ar1}
\title{Generate a TL-AR(1) time series}
\usage{
gen_ar1(n, phi)
}
\arguments{
\item{n}{the length of the desired time series}

\item{phi}{the AR(1) parameter}
}
\value{
a \code{vector}. The length-\code{n} times series with AR(1) parameter \code{phi}.
}
\description{
This function generates a TL-AR(1) time series of length \code{n}.
}
\examples{
out <- gen_ar1(1000, phi = 0.2)
}
\references{
\insertRef{mhatre2024arma}{tpdmethods}
}
