% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{gen_arma}
\alias{gen_arma}
\title{Generate a TL-ARMA(p, q) time series}
\usage{
gen_arma(n, phi, theta)
}
\arguments{
\item{n}{the length of the desired time series}

\item{phi}{a p-\code{vector} of the auto-regressive (AR) parameters where p is the
order of the AR component of the TL-ARMA(p,q) process.}

\item{theta}{a q-\code{vector} of the moving average (MA) parameters where q is the
order of the MA component of the TL-ARMA(p,q) process.}
}
\value{
a \code{vector}. The length-\code{n} times series with AR parameters \code{phi}
and MA parameters \code{theta}.
}
\description{
This function generates a TL-ARMA(p, q) time series of length \code{n}.
}
\examples{
out <- gen_arma(1000, phi = c(0.3), theta = c(-0.1, 0.4, 0.2))
}
\references{
\insertRef{mhatre2024arma}{tpdmethods}
}
