% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{gen_arma11}
\alias{gen_arma11}
\title{Generate a TLETS-ARMA(1, 1) time series}
\usage{
gen_arma11(n, phi, theta)
}
\arguments{
\item{n}{the length of the desired time series}

\item{phi}{the AR(1) parameter}

\item{theta}{the MA(1) parameter}
}
\value{
a \code{vector}. The length-\code{n} times series with AR(1) parameter \code{phi}
and MA(1) parameter \code{theta}.
}
\description{
This function generates a TLETS-ARMA(1,1) time series of length \code{n}.
}
\examples{
out <- gen_arma11(1000, phi = 0.3, theta = -0.1)
}
\references{
\insertRef{mhatre2024arma}{tpdmethods}
}
