% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{get_arma11_h}
\alias{get_arma11_h}
\title{Compute the TPD value for the lag-h pair from a TL-ARMA(1,1) model}
\usage{
get_arma11_h(params, h)
}
\arguments{
\item{params}{a vector of the form (\code{theta}, \code{phi}) where \code{theta} is the
TL-MA coefficient and \code{phi} is the TL-AR coefficient}

\item{h}{the lag for the desired TPD value}
}
\value{
a scalar TPD value
}
\description{
This function takes a \code{vector} pair of parameter values as input and outputs
the model TPDF for a single lag \code{h}. This function is called by
\code{arma11_tpdf} for each lag.
}
\examples{
out <- get_arma11_h(c(0.4, 0.1), h = 3)
}
\references{
\insertRef{mhatre2024arma}{tpdmethods}
}
