% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{get_eigen}
\alias{get_eigen}
\title{Get the eigenvalues and eigenvectors from a \code{matrix} of TPD values}
\usage{
get_eigen(tpdm, make_pd = TRUE)
}
\arguments{
\item{tpdm}{a square \code{matrix} of TPD values}

\item{make_pd}{change to \code{FALSE} if you do not want to find a PD matrix
that is close to the input matrix.}
}
\value{
a \code{list} with two components. The \code{values} are the eigenvalues of
the TPDM. The \code{vectors} are the eigenvectors of the TPDM.
}
\description{
This is a wrapper of the \code{eigen()} function that includes a step to find
the closest PD matrix to the input TPDM using the \code{nearPD()} function.
}
\examples{
set.seed(22)
myData <- matrix(rnorm(625), ncol = 5)
myTPDM <- tpd(myData)
out <- get_eigen(myTPDM)
}
