% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{get_pc_ts}
\alias{get_pc_ts}
\title{Compute the time series of a single principle component.}
\usage{
get_pc_ts(eigen_num, eigen_vecs, data)
}
\arguments{
\item{eigen_num}{the principal component to compute}

\item{eigen_vecs}{the \code{matrix} of eigenvectors.}

\item{data}{the full \code{matrix} of data.}
}
\value{
a \code{vector} of the principal components for each observation.
}
\description{
Compute the time series of a single principle component.
}
\examples{
myData <- matrix(rnorm(625), ncol = 5)
myTPDM <- tpd(myData)
myEigen <- get_eigen(myTPDM)
out <- get_pc_ts(1, eigen_vecs = myEigen$vectors, data = myData)
}
\references{
\insertRef{cooley_thibaud_2019}{tpdmethods}
}
