% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{get_pcs}
\alias{get_pcs}
\title{Get the principal components of a data vector}
\usage{
get_pcs(data_vec, eigenvecs)
}
\arguments{
\item{data_vec}{the data vector}

\item{eigenvecs}{the eigenvectors of the TPDM of the data vector}
}
\value{
a \code{vector} of principal components
}
\description{
Get the principal components of a data vector
}
\examples{
myData <- matrix(evd::rfrechet(625, shape = 2), ncol = 5)
myTPDM <- tpd(myData)
myEigen <- get_eigen(myTPDM)
myDatavec <- abs(evd::rfrechet(5, shape = 2))
out <- get_pcs(myDatavec, eigenvecs = myEigen$vectors)
}
\references{
\insertRef{cooley_thibaud_2019}{tpdmethods}
}
