% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{get_ts_plot}
\alias{get_ts_plot}
\title{Plot a time series of a single principle component}
\usage{
get_ts_plot(eigen_num, eigen_vecs, data, dates)
}
\arguments{
\item{eigen_num}{the principal component to plot.}

\item{eigen_vecs}{the \code{matrix} of eigenvectors.}

\item{data}{the full \code{matrix} of data.}

\item{dates}{the dates to put on the x-axis. If \code{dates == FALSE} no a default
x-axis will be used.}
}
\value{
a ggplot2 \code{plot} object
}
\description{
This function is a wrapper to compute and plot a time series for a single
principle component.
}
\examples{
myData <- matrix(rnorm(625), ncol = 5)
myTPDM <- tpd(myData)
myEigen <- get_eigen(myTPDM)
out <- get_ts_plot(eigen_num = 1,
                   eigen_vecs = myEigen$vectors,
                   data = myData,
                   dates = FALSE)
}
\references{
\insertRef{cooley_thibaud_2019}{tpdmethods}
}
