% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd.R
\name{hill_est}
\alias{hill_est}
\title{The Hill estimator for alpha}
\usage{
hill_est(x, k)
}
\arguments{
\item{x}{the \code{vector} of observations to be used to estimate alpha}

\item{k}{the number of large order statistics to be used}
}
\value{
a scalar value which is the point estimate for alpha
}
\description{
Estimates the index of regular variation (alpha) given a vector of
observations \code{x} and a number of large order statistics \code{k} to be used.
}
\examples{
hill_est(financial_data[,2], k = 300)
}
\references{
\insertRef{hill1975}{tpdmethods}
}
