% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{innovations}
\alias{innovations}
\title{Extremal analogue to the innovations algorithm}
\usage{
innovations(tpdf, max_q = 50)
}
\arguments{
\item{tpdf}{a \code{vector} of tpdf values where the first value in the vector is
the lag-0 tpd value.}

\item{max_q}{the largest order MA(q) that you want to fit. This could be your
desired order or some other large value to see when the estimates
stabilize.}
}
\value{
a \code{list} with two components. The first component \code{out[[1]]}  is a
\code{matrix} of innovations-estimated coefficients. The \code{n}th row contains
coefficients for the TL-MA(\code{n}). The second component \code{out[[2]]} is a
vector of \code{nu} values from the innovations algorith.
}
\description{
This function estimates TL-MA(q) coefficients using the extremes-analogue
to the innovations algorithm for model of order q = 1, ..., max_q. The
user inputs the estimated (or known) TPDF values and chooses a maximum
order to fit with the recursive algorithm.
}
\examples{
myTPD <- c(1, 0.4, 0.34, 0.2, 0.11, 0.05, 0.01)
out <- innovations(myTPD, max_q = 5)
}
\references{
\insertRef{mhatre2023innov}{tpdmethods}
}
