% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd.R
\name{is.tpd}
\alias{is.tpd}
\title{Check whether \code{x} is a \code{tpd} S3 object}
\usage{
is.tpd(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{TRUE} if \code{x} is a \code{numeric} \code{tpd} object, \code{FALSE} otherwise
}
\description{
Check whether \code{x} is a \code{tpd} S3 object
}
\examples{
my_tpd <- new_tpd(x = matrix(c(0.7, 1, 1, 0.7), nrow = 2), alpha = 2)
is.tpd(my_tpd)

is.tpd(1:5)
}
