% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{make_pd}
\alias{make_pd}
\title{Find a positive definite \code{matrix} that is close to the input \code{matrix}.}
\usage{
make_pd(x)
}
\arguments{
\item{x}{square \code{matrix} that may or may not be PD}
}
\value{
a \code{matrix} that is PD
}
\description{
This is just a wrapper for the \code{Matrix} function \code{nearPD()}
}
\examples{
myData <- matrix(evd::rfrechet(25, shape = 2), ncol = 5)
out <- make_pd(myData)
}
