% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{maq_tpdf}
\alias{maq_tpdf}
\title{Compute the TPD function from TL-MA(q) parameters}
\usage{
maq_tpdf(thetas, max_lag = 20)
}
\arguments{
\item{thetas}{the q-\code{vector} of parameters from a TL-MA(q) model}

\item{max_lag}{the TPDF is computed for lags 1 through \code{max_lag}
(default is 20)}
}
\value{
a \code{vector} of model TPDF values
}
\description{
This function takes a parameter vector \code{thetas} as input and returns the
model TPDF for lags 0 through \code{max_lag}
}
\examples{
out <- maq_tpdf(c(0.8, 0.4, 0.01), max_lag = 5)
}
\references{
\insertRef{mhatre2024arma}{tpdmethods}
}
