% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd.R
\name{new_tpd}
\alias{new_tpd}
\title{Constructor for the S3 class \code{tpd}.}
\usage{
new_tpd(x = double(), alpha = numeric(1))
}
\arguments{
\item{x}{the \code{numeric} values of the tpd object. A \code{vector} for a tpdf and a
\code{matrix} for a tpdm.}

\item{alpha}{the tail index of the underlying data.}
}
\value{
a \code{tpd} S3 object
}
\description{
Constructor for the S3 class \code{tpd}.
}
\examples{
my_tpd <- new_tpd(x = matrix(c(0.7, 1, 1, 0.7), nrow = 2), alpha = 2)
}
