% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_eigen_map}
\alias{plot_eigen_map}
\title{Plot a single eigenvector.}
\usage{
plot_eigen_map(
  eigen_num,
  eigen_vecs,
  eigen_vals,
  lat,
  lon,
  same_scale = FALSE,
  scale_lims = c(-99, -99)
)
}
\arguments{
\item{eigen_num}{which eigenvector you want to plot}

\item{eigen_vecs}{the \code{matrix} of eigenvectors}

\item{eigen_vals}{the \code{vector} of eigenvalues}

\item{lat}{a \code{vector} of latitudes}

\item{lon}{a \code{vector} of longitudes}

\item{same_scale}{change to \code{TRUE} if you want to force several plots to
have the same scale. (default is \code{FALSE})}

\item{scale_lims}{a bivariate \code{vector} of the form (scale minimum, scale
maximum). (default is c(-99, -99))}
}
\value{
a ggplot2 \code{plot} object
}
\description{
This is a wrapper fuction to plot a map showing the value of a single
eigenvector using latitude and longitude.
Note that eigenvalues and eigenvectors are separated for easier
plotting of transformed-linear reconstructions.
}
\examples{
myData <- matrix(evd::rfrechet(625, shape = 2), ncol = 5)
myTPDM <- tpd(myData)
myEigen <- get_eigen(myTPDM)
plot1 <- plot_eigen_map(1,
                        eigen_vecs = myEigen$vectors,
                        eigen_vals = myEigen$values,
                        lat = 1:5,
                        lon = 1)
plot1
}
