% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_reconstruction}
\alias{plot_reconstruction}
\title{Create plots for storm reconstruction with differing numbers of eigenvectors}
\usage{
plot_reconstruction(
  data,
  storm_ind,
  dates,
  lat,
  lon,
  eigenvectors,
  eigenvalues,
  reconstruction_inds = c(1, 2, 5, 20, 100)
)
}
\arguments{
\item{data}{The data \code{matrix}}

\item{storm_ind}{The index of the large storm day}

\item{dates}{The \code{vector} of \code{dates}}

\item{lat}{the \code{vector} of latitudes}

\item{lon}{the \code{vector} of longitudes}

\item{eigenvectors}{the \code{matrix} of eigenvectors}

\item{eigenvalues}{the \code{vector} of eigenvalues}

\item{reconstruction_inds}{a \code{vector} of the number of eigenvectors you want
used in a partial reconstruction (e.g., c(1, 5) would create a partial
reconstruction with 1 eigenvector and a second with 5 eigenvectors).
(default is \code{c(1, 2, 5, 20, 100)}).}
}
\value{
a \code{list} of \code{ggplot2} objects
}
\description{
This function is a wrapper on the \code{plot_sf()} function which outputs a \code{list}
of \code{ggplot2} objects. The first plot is the historical storm. The
following plots are partial storm reconstructions using
\code{reconstruction_inds} eigenvectors.
}
\examples{
myData <- matrix(evd::rfrechet(600, shape = 2), ncol = 6)
myTPDM <- tpd(myData)
myEigen <- get_eigen(myTPDM)
out <- plot_reconstruction(data = myData,
                           storm_ind = 2,
                           dates = 1:125,
                           lat = 1:2,
                           lon = 1:3,
                           eigenvectors = myEigen$vectors,
                           eigenvalues = myEigen$values,
                           reconstruction_inds = c(1,2,3))
cowplot::plot_grid(plotlist = out, nrow = 2)
}
