% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_scree}
\alias{plot_scree}
\title{Create a scree plot for TPDM eigenvalues}
\usage{
plot_scree(eigen_vals, max_ind = 10, normalize = TRUE)
}
\arguments{
\item{eigen_vals}{The \code{vector} of eigenvalues.}

\item{max_ind}{The index of the last eigenvalue you want included in the
plot. This is necessary when you have many eigenvalues and do not want to
plot them all. (default is 10)}

\item{normalize}{Change to \code{FALSE} if you do not want to normalize the
eigenvalues so that they sum to 1. (default is \code{TRUE})}
}
\value{
a \code{ggplot2} plot object
}
\description{
This function is a \code{ggplot2} wrapper which creates a scree plot (elbow plot)
for the first \code{max_ind} eigenvalues from an eigen decomposition.
}
\examples{
myData <- matrix(evd::rfrechet(625, shape = 2), ncol = 5)
myTPDM <- tpd(myData)
myEigen <- get_eigen(myTPDM)
out <- plot_scree(myEigen$values, max_ind = 3, normalize = FALSE)
}
