% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_sf}
\alias{plot_sf}
\title{Plot a \code{vector} of data using latitude and longitude}
\usage{
plot_sf(data, plot_title, lat, lon)
}
\arguments{
\item{data}{the \code{vector} of data to be plotted}

\item{plot_title}{the name of the plot}

\item{lat}{a \code{vector} of latitudes}

\item{lon}{a \code{vector} of longitudes}
}
\value{
a ggplot2 \code{plot} object
}
\description{
This is a wrapper function for creating and plotting an \code{sf} object. We use
it to plot partial reconstructions created with \code{reconstruct_data()}.
}
\examples{
myData <- rnorm(25)
out <- plot_sf(myData, plot_title = "Example Plot", lat = 1:5, lon = 1:5)
}
