% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{reconstruct_data}
\alias{reconstruct_data}
\title{Reconstruct a data vector with a subset of the eigenvectors.}
\usage{
reconstruct_data(pcs, eigenvecs, num_eigen)
}
\arguments{
\item{pcs}{the principal components of the data vector to be reconstructed.
Often this is the output of \code{get_pcs()}.}

\item{eigenvecs}{the eigenvectors of the TPDM of the data vector to be
reconstructed.}

\item{num_eigen}{the number of eigenvectors to be used in the partial
reconstruction}
}
\value{
a \code{vector} that contains the values of the partially reconstructed
data vector.
}
\description{
This function allows for the partial reconstruction of a data vector (for
example, of a big storm) using the \code{num_eigen} largest eigenvectors.
}
\examples{
myData <- matrix(evd::rfrechet(625, shape = 2), ncol = 5)
myTPDM <- tpd(myData)
myEigen <- get_eigen(myTPDM)
myDatavec <- abs(evd::rfrechet(5, shape = 2))
myPCs <- get_pcs(myDatavec, eigenvecs = myEigen$vectors)
out <- reconstruct_data(myPCs, eigenvecs = myEigen$vectors, num_eigen = 3)
}
\references{
\insertRef{cooley_thibaud_2019}{tpdmethods}
}
