% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{root_check}
\alias{root_check}
\title{Returns the modulus of the root of a polynomial that is closest to the origin.}
\usage{
root_check(phi)
}
\arguments{
\item{phi}{the parameter vector of the AR polynomial to be checked.}
}
\value{
the scalar modulus of the smallest root
}
\description{
Returns the modulus of the root of a polynomial that is closest to the origin.
}
\examples{
root_check(phi = c(0.8, 0.11, 0.1))
root_check(phi = c(0.8, 0.01, 0.1))
}
