% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_lin_operations.R
\name{tadd}
\alias{tadd}
\title{Transformed-linear addition}
\usage{
tadd(a, b)
}
\arguments{
\item{a}{a non-negative number}

\item{b}{a non-negative number}
}
\value{
a non-negative number
}
\description{
This function performes transformed-linear addition.
}
\examples{
tadd(1.2, 0.05)
tadd(125, 23.1)
}
\references{
\insertRef{cooley_thibaud_2019}{tpdmethods}
}
