% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_lin_operations.R
\name{tmult}
\alias{tmult}
\title{Transformed-linear scalar multiplication}
\usage{
tmult(a, b)
}
\arguments{
\item{a}{scalar}

\item{b}{\code{vector} of non-negative numbers}
}
\value{
a \code{vector}
}
\description{
This function performs transformed-linear scalar multiplication between for
scalar \code{a} and vector \code{b}
}
\examples{
tmult(0.5, c(1.2, 2.5, 0.3, 25.1))
}
\references{
\insertRef{cooley_thibaud_2019}{tpdmethods}
}
