% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd.R
\name{tpd.est}
\alias{tpd.est}
\title{Estimate the Tail Pairwise Dependence from large points}
\usage{
tpd.est(data)
}
\arguments{
\item{data}{a bivariate \code{matrix} of large RV(2) points.}
}
\value{
the estimated TPD value
}
\description{
This function estimates the TPD from a bivariate \code{matrix} of large regularly
varying (alpha = 2) points.
}
\examples{
myData <- matrix(evd::rfrechet(1000, shape = 2), ncol = 2)
radii <- sqrt(rowSums(myData^2))
largeData <- myData[order(radii)[450:500], ]
out <- tpd.est(largeData)
}
\references{
\insertRef{cooley_thibaud_2019}{tpdmethods}
}
