% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd.R
\name{tpd.once}
\alias{tpd.once}
\title{Get the Tail Pairwise Dependence of one pair of variables}
\usage{
tpd.once(
  data,
  radial_quantile = 0.975,
  radial_thresh = 0,
  get_large = TRUE,
  trans_marginal = TRUE,
  gpd_scale = -99,
  gpd_shape = -99,
  marginal_thresh = 0.975
)
}
\arguments{
\item{data}{a bivariate \code{matrix}}

\item{radial_quantile}{the quantile cutoff for computing the TPD. Points with
radial quantiles above this value will be used to estimate the TPD.
(default is 0.975)}

\item{radial_thresh}{instead of specifying the \code{radial_quantile} users can
specify a threshold such that points with radii larger than the
\code{radial_thresh} are included in the TPD estimation. \code{radial_quantile} is
ignored if this is greater than zero. (default is 0)}

\item{get_large}{change to \code{FALSE} if you have already removed all small
points. If \code{true} the function will remove points below the with radii
below the \code{radial_quantile} so that only large points are used in the
TPD computation. (default is \code{TRUE})}

\item{trans_marginal}{change to \code{FALSE} if the margins are already
Fr\'echet(2). If \code{TRUE} the function will transform marginal distributions
to be Fr\'echet(2). (default is \code{TRUE})}

\item{gpd_scale}{fix the scale at your desired scale value during the
marginal transformation. This is only used if \code{trans_marginal == TRUE}.
If no value is input then it will be estimated. (default is -99)}

\item{gpd_shape}{fix the shape at your desired shape value. If no value is
input then it will be estimated. Note that gpd shape is 1/alpha so if you
have generated data with alpha=2 you should enter 1/2. This is only used
if \code{trans_marginal == TRUE}. If no value is input then it will be
estimated. (default is -99)}

\item{marginal_thresh}{the quantile to use as a cutoff between the ECDF and
GPD components in the marginal transformation. (default is 0.975)}
}
\value{
the estimated TPD value
}
\description{
This function takes a bivariate matrix as input and estimates the TPD. If
necessary it transforms the marginal and gets the large points first.
}
\examples{
myData <- matrix(rnorm(10000), ncol = 2)
out1 <- tpd(myData, radial_quantile = 0.95)
}
