% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{transform_marginal}
\alias{transform_marginal}
\title{Transform the marginal distribution to be Fr\'echet(2)}
\usage{
transform_marginal(
  x,
  q = 0.975,
  use_gpd = TRUE,
  fix_gpd_params = FALSE,
  gpd_scale = -99,
  gpd_shape = -99
)
}
\arguments{
\item{x}{univariate data (\code{vector}) to be transformed}

\item{q}{the quantile to use as a cutoff between the ECDF and GPD components}

\item{use_gpd}{change to \code{FALSE} if you want to use a rank-transformation
only. When \code{TRUE} a GPD is used to estimate the upper tail. (default is
\code{TRUE})}

\item{gpd_scale}{fix the scale at your desired scale value. If no value is
input then it will be estimated. (default is -99)}

\item{gpd_shape}{fix the shape at your desired shape value. If no value is
input then it will be estimated. Note that gpd shape is 1/alpha so if you
have generated data with alpha=2 you should enter 1/2. (default is -99)}
}
\value{
a \code{vector} of data on Fr\'echet(2) margins
}
\description{
This function transforms the marginal distribution of a dataset. It is used on
one margin at a time. It uses a GPD above a quantile \code{q} and the ECDF below.
}
\examples{
myData <- rnorm(1000)
out <- transform_marginal(myData, q = 0.95)
}
