% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd.R
\name{validate_tpd}
\alias{validate_tpd}
\title{Validator for the S3 class \code{tpd}}
\usage{
validate_tpd(x)
}
\arguments{
\item{x}{the \code{tpd} object to be validated}
}
\value{
either an ERROR or the \code{tpd} object
}
\description{
Validator for the S3 class \code{tpd}
}
\examples{
my_tpd <- new_tpd(x = matrix(c(0.7, 1, 1, 0.7), nrow = 2), alpha = 2)
my_validated_tpd <- validate_tpd(my_tpd)
}
