% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{financial_data}
\alias{financial_data}
\title{Financial Data from the Kenneth R. French Data Library.}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\description{
These data are the losses in 30 different sectors of the economy (e.g.,
food, oil, and health) from 1970 through 2023. A \code{matrix} with 13599 rows
(days financial data were recorded) and 31 columns (first column is date,
the rest are the sectors).
}
\references{
\insertRef{kiriliouk2022}{tpdmethods}
}
