% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpd.R
\name{tpd}
\alias{tpd}
\title{Compute the Tail Pairwise Dependence (TPD) of an \code{R} object}
\usage{
tpd(
  data,
  radial_quantile = 0.975,
  max_lag = 50,
  radial_thresh = 0,
  get_large = TRUE,
  trans_marginal = TRUE,
  marginal_thresh = 0.975,
  matrix_as_seasons = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{as a \code{matrix}, \code{vector}, or \code{list}}

\item{radial_quantile}{the quantile cutoff for computing the TPD. Points with
radial quantiles above this value will be used to estimate the TPD.
(default is 0.975)}

\item{max_lag}{the maximum lag-value to compute the TPD for. Only used with
\code{vector} inputs. (default is 50)}

\item{radial_thresh}{instead of specifying the \code{radial_quantile} users can
specify a threshold such that points with radii larger than the
\code{radial_thresh} are included in the TPD estimation. \code{radial_quantile} is
ignored if this is greater than zero. (default is 0)}

\item{get_large}{change to \code{FALSE} if you have already removed all small
points. If \code{true} the function will remove points below the with radii
below the \code{radial_quantile} so that only large points are used in the
TPD computation. (default is \code{TRUE})}

\item{trans_marginal}{change to \code{FALSE} if the margins are already
Fr\'echet(2). If \code{TRUE} the function will transform marginal distributions
to be Fr\'echet(2). (default is \code{TRUE})}

\item{marginal_thresh}{the quantile to use as a cutoff between the ECDF and
GPD components in the marginal transformation. (default is 0.975)}

\item{matrix_as_seasons}{change to \code{TRUE} if your time series data are stored
as a matrix with columns representing plausibly iid replicates (e.g.,
seasons). This protects against lagging across seasons. (default is
\code{FALSE})}

\item{verbose}{change to \code{FALSE} if you do not want to see comments.
(default is \code{TRUE})}
}
\value{
the estimated tpd
}
\description{
This function puts the users input into a form recognizable by simpler
functions which transform the marginal distribution, extract the large
points, and compute the TPD value. It currently recognizes \code{matrices},
\code{vectors} as time series, and \code{lists} of bivariate data. \code{Matrices} can
be bivariate data, multivariate data, or time series data split into
seasons (use \code{matrix_as_seasons} = \code{TRUE}).
}
\examples{
myData <- matrix(rnorm(10000), ncol = 2)
out1 <- tpd(myData, radial_quantile = 0.95)

myMatrix <- matrix(rnorm(10000), ncol = 5)
out2 <- tpd(myMatrix, radial_thresh = 2)

myVector <- gen_ar1(1000, phi = 0.6)
out3 <- tpd(myVector, trans_marginal = FALSE)
}
\references{
\insertRef{cooley_thibaud_2019}{tpdmethods}
}
