% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{transform_marginal}
\alias{transform_marginal}
\title{Transform the marginal distribution to be Fr\'echet(2)}
\usage{
transform_marginal(
  x,
  q = 0.975,
  use_gpd = TRUE,
  fix_gpd_params = FALSE,
  gpd_scale = -99,
  gpd_shape = -99
)
}
\arguments{
\item{x}{univariate data (\code{vector}) to be transformed}

\item{q}{the quantile to use as a cutoff between the ECDF and GPD components}

\item{use_gpd}{change to \code{FALSE} if you want to use a rank-transformation
only. When \code{TRUE} a GPD is used to estimate the upper tail. (default is
\code{TRUE})}

\item{fix_gpd_params}{change to \code{TRUE} if you want to fix either the scale
or the shape of the GPD used to fit the upper tail of the marginal
distribution. (default is \code{FALSE})}

\item{gpd_scale}{set at your desired scale value. Only used if
\code{fix_gpd_params = TRUE}. (default is -99)}

\item{gpd_shape}{set at your desired shape value. Only used if
\code{fix_gpd_params = TRUE}. Note that gpd shape is 1/alpha. (default is -99)}
}
\value{
a \code{vector} of data on Fr\'echet(2) margins
}
\description{
This function transforms the marginal distribution of a dataset. It is used on
one margin at a time. It uses a GPD above a quantile q and the ECDF below.
}
\examples{
myData <- rnorm(1000)
out <- transform_marginal(myData, q = 0.95)
}
