% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlets.R
\name{transform_marginal}
\alias{transform_marginal}
\title{Transform the marginal distribution to be Fr\'echet(2)}
\usage{
transform_marginal(x, q = 0.975)
}
\arguments{
\item{x}{univariate data (\code{vector}) to be transformed}

\item{q}{the quantile to use as a cutoff between the ECDF and GPD components}
}
\value{
a \code{vector} of data on Fr\'echet(2) margins
}
\description{
This function transforms the marginal distribution of a dataset. It is used on
one margin at a time. It uses a GPD above a quantile q and the ECDF below.
}
\examples{
myData <- rnorm(1000)
out <- transform_marginal(myData, q = 0.95)
}
